requirements.txt
psycopg2
urllib3
python-dotenv
selenium
scrapy
